Extracted Yaml extensions out of Configuration.swift
