+++ Source/SwiftLintFramework/Models/Configuration.swift
@@ -12,16 +12,7 @@ import Yaml
 
 private let fileManager = NSFileManager.defaultManager()
 
-extension Yaml {
-    var arrayOfStrings: [Swift.String]? {
-        return array?.flatMap { $0.string } ?? string.map { [$0] }
-    }
-    var arrayOfInts: [Swift.Int]? {
-        return array?.flatMap { $0.int } ?? int.map { [$0] }
-    }
-}
-
-public struct Configuration: Equatable {
+public struct Configuration {
     public let disabledRules: [String] // disabled_rules
     public let included: [String]      // included
     public let excluded: [String]      // excluded
